requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
pytest==7.4.2
python-dotenv==1.0.0
werkzeug==2.2.2
